refactor: move MixerResourceDeserializer to common module
