+++ istio-common/src/main/java/me/snowdrop/istio/api/internal/MixerResourceDeserializer.java
@@ -11,7 +11,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package me.snowdrop.istio.api.policy.v1beta1;
+package me.snowdrop.istio.api.internal;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.ObjectCodec;
